Refactor: file directories
