+++ src/components/atoms/Button/index.ts
@@ -1,3 +1,4 @@
 import { HeaderButton } from "./HeaderButton"
+import { IconLinkButton } from "./IconLinkButton"
 
-export { HeaderButton }
+export { HeaderButton, IconLinkButton }
