+++ src/components/atoms/Typography/index.ts
@@ -1,0 +1,19 @@
+import { TitleTypography } from "./Title/TitleTypography"
+
+import { Subtitle1Typography } from "./Subtitle/Subtitle1Typography"
+import { Subtitle2Typography } from "./Subtitle/Subtitle2Typography"
+import { Subtitle3Typography } from "./Subtitle/Subtitle3Typography"
+
+import { H1Typography } from "./Heading/H1Typography"
+import { H2Typography } from "./Heading/H2Typography"
+import { H3Typography } from "./Heading/H3Typography"
+import { H4Typography } from "./Heading/H4Typography"
+
+import { P1Typography } from "./Paragraph/P1Typography"
+import { P2Typography } from "./Paragraph/P2Typography"
+import { P3Typography } from "./Paragraph/P3Typography"
+
+export { TitleTypography }
+export { Subtitle1Typography, Subtitle2Typography, Subtitle3Typography }
+export { P1Typography, P2Typography, P3Typography }
+export { H1Typography, H2Typography, H3Typography, H4Typography }
