+++ src/components/layouts/Footer/index.ts
@@ -1,0 +1,3 @@
+import { ResponsibleFooterLayout } from "./ResponsibleFooterLayout"
+
+export { ResponsibleFooterLayout }
