+++ src/components/layouts/Header/Header.tsx
@@ -3,8 +3,7 @@ import {
   useTheme,
 } from "@mui/material"
 import React from "react"
-import { PCHeaderLayout } from "./PC/PCHeaderLayout"
-import { SPHeaderLayout } from "./SP/SPHeaderLayout"
+import { PCHeaderLayout, SPHeaderLayout } from "@/components/layouts/Header"
 
 export const Header = () => {
   const theme = useTheme()
