+++ src/components/layouts/Header/index.ts
@@ -1,0 +1,4 @@
+import { PCHeaderLayout } from "./PC/PCHeaderLayout"
+import { SPHeaderLayout } from "./SP/SPHeaderLayout"
+
+export { PCHeaderLayout, SPHeaderLayout }
