+++ src/components/molecules/MenuList/index.ts
@@ -1,0 +1,4 @@
+import { PCMenuList } from "./PCMenuList/PCMenuList"
+import { SPMenuList } from "./SPMenuList/SPMenuList"
+
+export { PCMenuList, SPMenuList }
